backend/requirements.txt
vertexai
aiosmtpd
requests
psycopg2-binary
flask